Syntax, validation, and partial execution for bag-o-stacks
Specify stack types as a sequence of value types received by the stack when it
is switched to, the last of which must be a reference to a defined stack type.
At runtime, that last stack reference value will be the handle to the previously
active stack or will be null if the previous stack was retired. Subtyping
between stack types is contravariant in all of the sent values.

Also specify the syntax and validation of the following instructions:

 - `stack.new` and `stack.new_ref`: Allocate a new stack that will call the
 provided function or function reference when it is switched to. Takes as
 operands a prefix of the function arguments, leaving the rest to be provided
 when the stack is switched to.

 - `stack.switch`: Switch to the provided stack. Takes as operands the values
 the stack expects to receive except for the final stack reference, which is
 created as part of the execution of this instruction.

 - `stack.switch_retire`: Switch to the provided stack as above, but sends it a
 null stack reference, allowing the previous stack resources to be freed. Can
 only be used with stack reference types that receive a nullable stack
 reference.

 - `stack.new_switch` and `stack.new_switch_ref`: Compositions of
 `stack.new` (or `stack.new_ref`) and `stack.switch` that may allow for better
 optimizations in engines.

Runtime semantics are partially worked out except for the interesting machinery
for allocating and switching to stacks.